M3 progress
Created the login screen.
This was done by changing the activity_main.xml.
Currently, very basic, has input fields, a background place holder, the
login button, and the register button.

Currently the login button is able to pass the input from the user to
the linked activity.

I also created a login clicked activity so that I could test to make
sure that the correct inputs were being passed.

Also, I added a background place holder image to the drawable-ldpi
folder. This can be removed later if need be, however, for now it is in
use until a better alternative is found.