Alterations
Changed LoginHandler to work with db and removed/commented out
references to the hashtable. Added most of the functionality required to
the DBHandler. Still have to fix the transaction table, but overall it
is coming along nicely. Added a getName methad to Account, so that I
could found out what the account name is simply from the object, it's
required by one of my DBHandler methods. Added a method call in the
addAccount function of the AccountManager class to make sure the account
is added to the DB that instant. Commented out all the
.readTable/.writeTable references I  could find in AdminActivity and
WelcomeContainer.